+++ examples/complete/outputs.tf
@@ -1,6 +1,1 @@
-output "mysql" {
-  value = [
-    module.rds_instance_mysql,
-  ]
-  description = "Output values for the module"
-}
+# Empty