+++ examples/mariadb/locals.tf
@@ -1,7 +1,13 @@
 locals {
-  name             = "randominstancemariadb"
-  cidr_block       = "172.16.0.0/16"
-  database_subnets = cidrsubnets(local.cidr_block, 8, 8, 8)
-  environment      = "examples"
-  azs              = flatten(data.aws_availability_zones.available.names)
+  name                      = "exampleinstancemariadb"
+  vpc_id                    = data.aws_vpc.supporting.id
+  database_subnets          = local.database_subnet_id
+  supporting_resources_name = "terraform-aws-rds"
+  database_subnet_id = [
+    for s in data.aws_subnet.database : s.id
+  ]
+  tags = {
+    Name        = local.name
+    Environment = "examples"
+  }
 }