+++ examples/mssql/main.tf
@@ -1,20 +1,6 @@
 #######################
 #  MsSQL Engine example
 #######################
-module "vpc" {
-  source               = "git::https://github.com/boldlink/terraform-aws-vpc.git?ref=2.0.3"
-  cidr_block           = local.cidr_block
-  name                 = local.name
-  enable_dns_support   = true
-  enable_dns_hostnames = true
-  account              = data.aws_caller_identity.current.account_id
-  region               = data.aws_region.current.name
-
-  ## database Subnets
-  database_subnets   = local.database_subnets
-  availability_zones = local.azs
-}
-
 resource "random_string" "rds_usr" {
   length  = 5
   special = false
@@ -28,16 +14,17 @@ resource "random_password" "rds_pwd" {
 }
 
 module "rds_instance_mssql" {
+  #checkov:skip=CKV_AWS_157: "Ensure that RDS instances have Multi-AZ enabled"
+  #checkov:skip=CKV_AWS_16: "Ensure all data stored in the RDS is securely encrypted at rest"
   source                          = "../../"
   engine                          = "sqlserver-ee"
   allocated_storage               = 30
   engine_version                  = "15.00.4153.1.v1"
   instance_class                  = "db.t3.xlarge"
-  subnet_ids                      = flatten(module.vpc.database_subnet_id)
+  subnet_ids                      = local.database_subnets
   name                            = local.name
   username                        = random_string.rds_usr.result
   password                        = random_password.rds_pwd.result
-  environment                     = local.environment
   port                            = 1433
   enabled_cloudwatch_logs_exports = ["agent", "error"]
   create_security_group           = true
@@ -45,14 +32,12 @@ module "rds_instance_mssql" {
   monitoring_interval             = 30
   create_option_group             = true
   deletion_protection             = false
-  vpc_id                          = module.vpc.id
+  vpc_id                          = local.vpc_id
   assume_role_policy              = data.aws_iam_policy_document.monitoring.json
   policy_arn                      = "arn:${data.aws_partition.current.partition}:iam::aws:policy/service-role/AmazonRDSEnhancedMonitoringRole"
   major_engine_version            = "15.00"
   timezone                        = "UTC"
   license_model                   = "license-included"
   storage_encrypted               = false
-  other_tags = {
-    "cost_center" = "random"
-  }
+  tags                            = local.tags
 }