+++ examples/mssql/outputs.tf
@@ -1,6 +1,1 @@
-output "address" {
-  value = [
-    module.rds_instance_mssql,
-  ]
-  description = "Values for resources created by the module"
-}
+# Empty